+++ keyboards/dcpedit/redherring/keymaps/default/rules.mk
@@ -1,2 +1,1 @@
-VIA_ENABLE = yes
 ENCODER_MAP_ENABLE = yes
